// docs/一期/archive/cli copy/src/config/project/types.ts
export interface CustomPropertyDescriptor {
  key: string;
  name?: string;
  type: string;
  defaultValue?: string;
  remark?: string;
  candidateItems?: string[];
}